cloud-functions/tag-bq-table/requirements.txt
functions-framework==3.*
google-cloud-logging
google-cloud-datacatalog